backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
asyncpg
google-auth
google-genai
Pillow
